# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "description": "A simple e-commerce API for managing products and orders with inventory management",
    "version": "1.0.0",
    "contact": {
      "name": "E-Commerce Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "http://api.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns the health status of the API",
        "tags": ["Health"],
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "all good"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List all products",
        "description": "Retrieve a list of all available products with pricing and stock information",
        "tags": ["Products"],
        "operationId": "listProducts",
        "responses": {
          "200": {
            "description": "List of products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "description": "Place a new order with products. Validates stock availability and creates order items within a transaction.",
        "tags": ["Orders"],
        "operationId": "placeOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - missing or invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - insufficient product stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "required": ["id", "name", "price_in_centers", "quantity", "created_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique product identifier"
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Laptop"
          },
          "price_in_centers": {
            "type": "integer",
            "format": "int32",
            "description": "Price in cents (divide by 100 for dollars)",
            "example": 99999
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "description": "Available stock quantity",
            "example": 10
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product creation timestamp"
          }
        }
      },
      "Order": {
        "type": "object",
        "required": ["id", "customer_id", "created_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique order identifier"
          },
          "customer_id": {
            "type": "integer",
            "format": "int64",
            "description": "Customer ID",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order creation timestamp"
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "required": ["product_id", "quantity"],
        "properties": {
          "product_id": {
            "type": "integer",
            "format": "int64",
            "description": "Product ID",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "description": "Quantity to order",
            "example": 2
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": ["customer_id", "items"],
        "properties": {
          "customer_id": {
            "type": "integer",
            "format": "int64",
            "description": "Customer ID",
            "example": 1
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "description": "List of items to order",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      }
    }
  }
}